Undid a few things I wasn't sure about before merging
Can re-do later, just wanted to grab the changes I'm sure of with proper credit
in the interim

+++ tests/parallel.hpp
@@ -44,12 +44,12 @@ TEST(ParallelTest, InterleavedAscending) {
 	system("mkdir -p thread-logs");
 
 	for (uint_fast16_t i = 0; i < PARALLEL_RERUNS; ++i) {
-		fprintf(log_stream, "Run %d\n", (unsigned)i+1);
+		fprintf(log_stream, "Run %d\n", i+1);
 		root = 0;
 		mem_reset_all(memory);
-		sprintf(strbuf, "thread-logs/int-asc_run_%04d_%s.log", (unsigned)i, "odd");
+		sprintf(strbuf, "thread-logs/int-asc_run_%04d_%s.log", i, "odd");
 		odd_args.log_stream = fopen(strbuf, "w");
-		sprintf(strbuf, "thread-logs/int-asc_run_%04d_%s.log", (unsigned)i, "even");
+		sprintf(strbuf, "thread-logs/int-asc_run_%04d_%s.log", i, "even");
 		even_args.log_stream = fopen(strbuf, "w");
 
 		pthread_create(&thread_even, NULL, stride_insert, (void*) &even_args);
@@ -96,12 +96,12 @@ TEST(ParallelTest, InterleavedDescending) {
 	system("mkdir -p thread-logs");
 
 	for (uint_fast16_t i = 0; i < PARALLEL_RERUNS; ++i) {
-		fprintf(log_stream, "Run %d\n", (unsigned)i+1);
+		fprintf(log_stream, "Run %d\n", i+1);
 		root = 0;
 		mem_reset_all(memory);
-		sprintf(strbuf, "thread-logs/int-des_run_%04d_%s.log", (unsigned)i, "odd");
+		sprintf(strbuf, "thread-logs/int-des_run_%04d_%s.log", i, "odd");
 		odd_args.log_stream = fopen(strbuf, "w");
-		sprintf(strbuf, "thread-logs/int-des_run_%04d_%s.log", (unsigned)i, "even");
+		sprintf(strbuf, "thread-logs/int-des_run_%04d_%s.log", i, "even");
 		even_args.log_stream = fopen(strbuf, "w");
 
 		pthread_create(&thread_even, NULL, stride_insert, (void*) &even_args);
@@ -146,12 +146,12 @@ TEST(ParallelTest, CrossfadeInsert) {
 	system("mkdir -p thread-logs");
 
 	for (uint_fast16_t i = 0; i < PARALLEL_RERUNS; ++i) {
-		fprintf(log_stream, "Run %d\n", (unsigned)i+1);
+		fprintf(log_stream, "Run %d\n", i+1);
 		root = 0;
 		mem_reset_all(memory);
-		sprintf(strbuf, "thread-logs/x-fade_run_%04d_%s.log", (unsigned)i, "odd");
+		sprintf(strbuf, "thread-logs/x-fade_run_%04d_%s.log", i, "odd");
 		odd_args.log_stream = fopen(strbuf, "w");
-		sprintf(strbuf, "thread-logs/x-fade_run_%04d_%s.log", (unsigned)i, "even");
+		sprintf(strbuf, "thread-logs/x-fade_run_%04d_%s.log", i, "even");
 		even_args.log_stream = fopen(strbuf, "w");
 
 		pthread_create(&thread_even, NULL, stride_insert, (void*) &even_args);
